added auth and profile management for osgi

+++ fitpal-springboot-app/src/main/java/com/fitpal/fitpalspringbootapp/models/User.java
@@ -39,7 +39,7 @@ public class User {
 
     private Double height;
 
-    private Integer activityLevel;
+    private Double activityLevel;
 
     private Integer weightGoal; // -500, 0, or 500
